EMS-54 Add base component for entities

+++ src/main/java/vn/com/tma/emsbackend/validation/validator/PortValidator.java
@@ -6,10 +6,10 @@ import vn.com.tma.emsbackend.validation.Port;
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 
-public class PortValidator implements ConstraintValidator<Port, Long> {
+public class PortValidator implements ConstraintValidator<Port, Integer> {
 
     @Override
-    public boolean isValid(Long value, ConstraintValidatorContext context) {
-            return (value > 0 && value < Constant.MAX_PORT_NUMBER);
+    public boolean isValid(Integer value, ConstraintValidatorContext context) {
+        return (value > 0 && value < Constant.MAX_PORT_NUMBER);
     }
 }
